Revert "using valhalla for request page"
