+++ valhalla/packages/jawn/src/index.ts
@@ -4,12 +4,7 @@ require("dotenv").config({
 
 import express from "express";
 import * as OpenApiValidator from "express-openapi-validator";
-import {
-  getRequests,
-  getRequestsCached,
-  withAuth,
-  withDB,
-} from "helicone-shared-ts";
+import { getRequests, withAuth, withDB } from "helicone-shared-ts";
 import morgan from "morgan";
 import { v4 as uuid } from "uuid";
 import { paths } from "./schema/types";
@@ -30,7 +25,6 @@ const errorHandler: ErrorRequestHandler = (
   res.status(500).send("Something broke!");
 };
 
-export const ENVIRONMENT = process.env.VERCEL_ENV ?? "development";
 const dirname = __dirname;
 
 const app = express();
@@ -46,57 +40,14 @@ app.use(express.urlencoded({ limit: "50mb" }));
 app.use(morgan("combined"));
 app.use(express.json()); // for parsing application/json
 
-app.use(errorHandler);
-const allowedOriginsEnv = {
-  production: [
-    /^https:\/\/helicone\.ai$/,
-    /^https:\/\/.*-helicone\.vercel\.app\/$/,
-    /^https:\/\/helicone\.vercel\.app\/$/,
-    /^https:\/\/helicone-git-valhalla-use-jawn-to-read-helicone\.vercel\.app$/,
-  ],
-  development: [/^http:\/\/localhost:3000$/, /^http:\/\/localhost:3001$/],
-};
-
-const corsForHelicone = (req: Request, res: Response, next: () => void) => {
-  const origin = req.get("Origin");
-  if (!origin) {
-    next();
-    return;
-  }
-
-  const allowedOrigins =
-    ENVIRONMENT === "development"
-      ? allowedOriginsEnv["development"]
-      : allowedOriginsEnv["production"];
-  const isAllowedOrigin = allowedOrigins.some((pattern) =>
-    pattern.test(origin)
-  );
-
-  if (isAllowedOrigin) {
-    // Set CORS headers
-    res.header("Access-Control-Allow-Origin", "*");
-    res.header("Access-Control-Allow-Methods", "GET,POST,PUT,PATCH,DELETE");
-    res.header(
-      "Access-Control-Allow-Headers",
-      "Content-Type, helicone-authorization"
-    );
-  } else {
-    res.header("info", `not allowed origin for ${ENVIRONMENT} environment :(`);
-  }
-  next();
-};
-
-app.use(corsForHelicone);
-app.options("*", (req, res) => {
-  res.sendStatus(200);
-});
-
 app.use(
   OpenApiValidator.middleware({
     apiSpec: process.env.OPENAPI_SCHEMA_FILE ?? `${dirname}/schema/openapi.yml`,
     validateRequests: true,
   })
 );
+
+app.use(errorHandler);
 
 app.post(
   "/v1/request/query",
@@ -107,32 +58,15 @@ app.post(
     console.log("body", body);
     const { filter, offset, limit, sort, isCached } = body;
 
-    const metrics = isCached
-      ? await getRequestsCached(
-          authParams.organizationId,
-          filter,
-          offset,
-          limit,
-          sort,
-          supabaseClient.client
-        )
-      : await getRequests(
-          authParams.organizationId,
-          filter,
-          offset,
-          limit,
-          sort,
-          supabaseClient.client
-        );
-    res
-      .header("Access-Control-Allow-Origin", "*")
-      .header("Access-Control-Allow-Methods", "GET,POST,PUT,PATCH,DELETE")
-      .header(
-        "Access-Control-Allow-Headers",
-        "Content-Type, helicone-authorization"
-      )
-      .status(metrics.error === null ? 200 : 500)
-      .json(metrics);
+    const metrics = await getRequests(
+      authParams.organizationId,
+      filter,
+      offset,
+      limit,
+      sort,
+      supabaseClient.client
+    );
+    res.status(metrics.error === null ? 200 : 500).json(metrics);
   })
 );
